+++ core/modules/catalog/store/category/actions.ts
@@ -212,7 +212,7 @@ const actions: ActionTree<CategoryState, RootState> = {
     let t0 = new Date().getTime()
 
     const precachedQuery = searchProductQuery
-    let productPromise = rootStore.dispatch('product/list', {
+    let productPromise = rootStore.dispatch('product/findProducts', {
       query: precachedQuery,
       start: current,
       size: perPage,
@@ -323,7 +323,7 @@ const actions: ActionTree<CategoryState, RootState> = {
 
     if (config.entities.twoStageCaching && config.entities.optimize && !isServer && !rootStore.state.twoStageCachingDisabled && !cacheOnly) { // second stage - request for caching entities; if cacheOnly set - the caching took place with the stage1 request!
       Logger.log('Using two stage caching for performance optimization - executing second stage product caching', 'category') // TODO: in this case we can pre-fetch products in advance getting more products than set by pageSize()
-      rootStore.dispatch('product/list', {
+      rootStore.dispatch('product/findProducts', {
         query: precachedQuery,
         start: current,
         size: perPage,