+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Reduce initial client-side bundle-size by lazy-loading `i18n` translations - @cewald (#4821)
+- Replaced deprecated action product/list call with product/findProducts (#4769)
 - Documentation examples of how to translate strings that contain variables or HTML tags - @weaves81 (#4991)
 
 ### Fixed